beneficiaryAddr move to web ui move /webservice/ to /dashboard/
