+++ dashboard/routes/node/status.go
@@ -1,7 +1,12 @@
 package node
 
 import (
+	"encoding/json"
 	"github.com/gin-gonic/gin"
+	"github.com/nknorg/nkn/node"
+	"github.com/nknorg/nkn/util/config"
+	"github.com/nknorg/nkn/util/log"
+	"net/http"
 )
 
 type NodeRouter struct {
@@ -9,9 +14,23 @@ type NodeRouter struct {
 
 func (nodeRouter *NodeRouter) Router(router *gin.RouterGroup) {
 	router.GET("/node/status", func(context *gin.Context) {
-		node, exists := context.Get("localNode")
+		var out map[string]interface{}
+
+		localNode, exists := context.Get("localNode")
+
 		if exists {
-			context.JSON(200, node)
+			buf, err := localNode.(*node.LocalNode).MarshalJSON()
+			if err != nil {
+				log.WebLog.Error(err)
+				context.AbortWithError(http.StatusInternalServerError, err)
+			}
+			err = json.Unmarshal(buf, &out)
+			if err != nil {
+				log.WebLog.Error(err)
+				context.AbortWithError(http.StatusInternalServerError, err)
+			}
+			out["beneficiaryAddr"] = config.Parameters.BeneficiaryAddr
+			context.JSON(200, out)
 			return
 		}
